chore: split achievement code
